# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Firecrawl API",
    "version": "1.0.0",
    "description": "API for interacting with Firecrawl services to perform web scraping and crawling tasks.",
    "contact": {
      "name": "Firecrawl Support",
      "url": "https://firecrawl.dev/support",
      "email": "support@firecrawl.dev"
    }
  },
  "servers": [
    {
      "url": "https://api.firecrawl.dev/v1"
    }
  ],
  "paths": {
...
    "/crawl/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "The ID of the crawl job",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Get the status of a crawl job",
        "operationId": "getCrawlStatus",
        "tags": ["Crawling"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlStatusResponseObj"
                }
              }
            }
          },
          "402": {
            "description": "Payment required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Payment required to access this resource."
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Request rate limit exceeded. Please wait and try again later."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "An unexpected error occurred on the server."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cancel a crawl job",
        "operationId": "cancelCrawl",
        "tags": ["Crawling"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful cancellation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["cancelled"],
                      "example": "cancelled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Crawl job not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Crawl job not found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "An unexpected error occurred on the server."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "CrawlStatusResponseObj": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "The current status of the crawl. Can be `scraping`, `completed`, or `failed`."
          },
          "total": {
            "type": "integer",
            "description": "The total number of pages that were attempted to be crawled."
          },
          "completed": {
            "type": "integer",
            "description": "The number of pages that have been successfully crawled."
          },
          "creditsUsed": {
            "type": "integer",
            "description": "The number of credits used for the crawl."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the crawl will expire."
          },
          "next": {
            "type": "string",
            "nullable": true,
            "description": "The URL to retrieve the next 10MB of data. Returned if the crawl is not completed or if the response is larger than 10MB."
          },
          "data": {
            "type": "array",
            "description": "The data of the crawl.",
            "items": {
              "type": "object",
              "properties": {
                "markdown": {
                  "type": "string"
                },
                "html": {
                  "type": "string",
                  "nullable": true,
                  "description": "HTML version of the content on page if `includeHtml`  is true"
                },
                "rawHtml": {
                  "type": "string",
                  "nullable": true,
                  "description": "Raw HTML content of the page if `includeRawHtml`  is true"
                },
                "links": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "List of links on the page if `includeLinks` is true"
                },
                "screenshot": {
                  "type": "string",
                  "nullable": true,
                  "description": "Screenshot of the page if `includeScreenshot` is true"
                },
                "metadata": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "language": {
                      "type": "string",
                      "nullable": true
                    },
                    "sourceURL": {
                      "type": "string",
                      "format": "uri"
                    },
                    "<any other metadata> ": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "integer",
                      "description": "The status code of the page"
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "description": "The error message of the page"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}